Rename `i18next` utils to `i18n`
